+++ packages/paraglide/paraglide-js-adapter-next/examples/app/src/lib/i18n.ts
@@ -6,6 +6,11 @@ export const { Link, middleware, useRouter, usePathname, redirect, permanentRedi
 	createI18n<AvailableLanguageTag>({
 		pathnames: {
 			"/about": m.about_path,
+			"/admin/[...rest]": {
+				en: "/admin/[...rest]",
+				de: "/administrator/[...rest]",
+				"de-CH": "/administrator/[...rest]",
+			},
 		},
 		prefix: "all",
 		exclude: ["/not-translated"], //makes sure that the /not-translated page is not translated