+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inlang/telemetry
 
+## 0.3.22
+
+### Patch Changes
+
+- Updated dependencies [bc9875d]
+  - @inlang/sdk@0.32.0
+
 ## 0.3.21
 
 ### Patch Changes